% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batch_file}
\alias{batch_file}
\title{Process the input file a batch of one batch}
\usage{
batch_file(input_file = "", batch_lines = 1e+07, handler = NULL,
  param_names = c("x", "i"), extra_params = list())
}
\arguments{
\item{input_file}{Filename need to process}

\item{batch_lines}{Batch lines to process the data, default 10000000}

\item{handler}{The function to process the data}

\item{param_names}{Hander function required parameter names}

\item{extra_params}{Extra paramemters pass to handler}
}
\description{
Process the input file a batch of one batch
}
\examples{
dat <- data.frame(a=1:100, b=1:100)
input_file <- tempfile()
write.table(dat, input_file, sep = "\\t", row.names = F, quote = F)
handler_fun <- function(x, i = 1) {
  return(x[i])
}
batch_file(input_file, 10, handler_fun)
}
