% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinpaint.R
\name{fusion2pp}
\alias{fusion2pp}
\title{Function to convert fusion data to ProteinPaint \url{https://pecan.stjude.org/pp} input format.}
\usage{
fusion2pp(input_data, format = "fusioncatcher", order = list(gene5 = 1,
  gene3 = 2))
}
\arguments{
\item{input_data}{A gene fusions data.frame need to be converted to ProteinPaint input.}

\item{format}{Point the input data format (annovar or others)}

\item{order}{Input data cols order (chr, start, end, trascript and aachange)}
}
\value{
A data frame
}
\description{
Function to convert fusion data to ProteinPaint \url{https://pecan.stjude.org/pp} input format.
}
\examples{
example <- NULL
}
