% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinpaint.R
\name{fusion2pp_meta}
\alias{fusion2pp_meta}
\title{Function to convert fusion data to ProteinPaint heatmap meta rows \url{https://pecan.stjude.org/pp} input format.}
\usage{
fusion2pp_meta(input_data, format = "fusioncatcher", order = list(sample =
  1, gene5 = 2, gene3 = 3))
}
\arguments{
\item{input_data}{A gene fusions data.frame need to be converted to ProteinPaint input (heatmap meta rows).}

\item{format}{Point the input data format (annovar or others)}

\item{order}{Input data cols order (chr, start, end, trascript and aachange)}
}
\value{
A data frame
}
\description{
Function to convert fusion data to ProteinPaint heatmap meta rows \url{https://pecan.stjude.org/pp} input format.
}
\examples{
example <- NULL
}
