% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{get_split_seqs}
\alias{get_split_seqs}
\title{Function to calculate the split regions by sections and total numbers}
\usage{
get_split_seqs(total_num, sections)
}
\arguments{
\item{total_num}{Total numbers need to be divided into n sections}

\item{sections}{Split section number (colnum)}
}
\description{
Function to calculate the split regions by sections and total numbers
}
\examples{
total_num <- 1000
sections <- 3
get_split_seqs(total_num, sections)
}
