% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinpaint_transefer.R
\name{mut2pp}
\alias{mut2pp}
\title{Function to convert mutation data to ProteinPaint \url{https://pecan.stjude.org/pp} input format.}
\usage{
mut2pp(input_data, input_type = "iseq", config_file = system.file("extdata",
  "config/proteinpaint.toml", package = "ngstk"))
}
\arguments{
\item{input_data}{A mutation data.frame need to be converted to ProteinPaint input.}

\item{input_type}{Point the input data format (iseq or others)}

\item{config_file}{ngstk ProteinPaint configuration file path, default is 
system.file('extdata', 'config/proteinpaint.toml', package = 'ngstk')}
}
\value{
A data frame
}
\description{
Function to convert mutation data to ProteinPaint \url{https://pecan.stjude.org/pp} input format.
}
\examples{
demo_file <- system.file('extdata', 
'demo/proteinpaint/mut2pp_iseq.txt', package = 'ngstk')
input_data <- read.table(demo_file, sep = '\\t', header = TRUE, stringsAsFactors = FALSE)
disease <- 'T-ALL'
input_data <- data.frame(input_data, disease)
input_data$disease <- as.character(input_data$disease)
mut2pp(input_data, input_type = 'iseq')
}
