% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngstk.R
\docType{package}
\name{ngstk}
\alias{ngstk}
\alias{ngstk-package}
\title{ngstk can be used to facilitate the analysis of NGS data, 
such as visualization, input data format convertion (for WEB service and others), 
database id convertion.}
\description{
ngstk can be used to facilitate the analysis of NGS data, 
such as visualization, input data format convertion (for WEB service and others), 
database id convertion.
}
\seealso{
Useful links:

\url{https://github.com/Miachol/ngstk}

Report bugs at \url{https://github.com/Miachol/ngstk/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
