% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{parse_args2}
\alias{parse_args2}
\title{Parse command line options.}
\usage{
parse_args2(...)
}
\arguments{
\item{...}{Parameters pass to \code{\link[optparse]{parse_args2}}}
}
\description{
Parse command line options.
}
\examples{
#example from vignette using positional arguments
option_list2 <- list(
  make_option(c('-n', '--add-numbers'), action='store_true', default=FALSE,
      help='Print line number at the beginning of each line [default]')
  )
parser <- opt_parser(usage = '\%prog [options] file', option_list=option_list2)

parse_args(parser, args = c('--add-numbers', 'example.txt'), positional_arguments = TRUE)

parse_args(parser, args = c('--add-numbers', 'example.txt'), positional_arguments = TRUE,
        convert_hyphens_to_underscores = TRUE)
parse_args2(parser, args = c('--add-numbers', 'example.txt'))
}
