% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{rbin}
\alias{rbin}
\title{Function to generate executable files for R package}
\usage{
rbin(pkgs, destdir = "~/.ngstk/bin")
}
\arguments{
\item{pkgs}{Names f R packages to generate executable files}

\item{destdir}{Destination directory to store the inst/bin files [~/.ngstk/bin]}

\item{...}{Params pass to \link{file.rename}.}
}
\description{
Function to generate executable files for R package
}
\examples{
rbin('ngstk', tempdir())
}
