% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_colors}
\alias{set_colors}
\title{Function to get a series defined theme colors}
\usage{
set_colors(theme = NULL, theme_config_file = NULL,
  show_all_themes = FALSE)
}
\arguments{
\item{theme}{Colors theme, e.g. default, red_blue}

\item{theme_config_file}{Theme configuration file, default is
system.file('extdata', 'config/theme.toml', package = 'ngstk')}

\item{show_all_themes}{Wheather show all avaliable colors theme, default is FALSE}
}
\value{
A character
}
\description{
Function to get a series defined theme colors
}
\examples{
red_blue <- set_colors('red_blue')
default <- set_colors('default')
colors <- set_colors(show_all_themes = TRUE)
}
