% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hander.R
\name{show_handers}
\alias{show_handers}
\title{Function to show all avaliabe handler function}
\usage{
show_handers(hander_lib = "default_handers", show_all_funs = TRUE,
  show_code = NULL, show_description = FALSE,
  hander_confg_file = system.file("extdata", "config/hander.toml", package =
  "ngstk"))
}
\arguments{
\item{hander_lib}{hander lib name}

\item{show_all_funs}{Default is TRUE and to show all functions in the hander_lib}

\item{show_code}{Default is NULL, select hander you want to see its source code}

\item{show_description}{Default is FALSE}

\item{hander_confg_file}{ngstk hander configuration file path, default is
system.file('extdata', 'config/hander.toml', package = 'ngstk')}
}
\description{
Function to show all avaliabe handler function
}
\examples{
show_handers(show_description = TRUE)
show_handers(show_description = FALSE, show_all_funs = FALSE, 
            show_code = 'handler_na_replace')
}
