% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{show_handlers}
\alias{show_handlers}
\title{Function to show all avaliabe handler function}
\usage{
show_handlers(handler_lib = "default_handlers", show_all_funs = TRUE,
  show_code = NULL, show_description = FALSE,
  handler_confg_file = system.file("extdata", "config/handler.toml",
  package = "ngstk"))
}
\arguments{
\item{handler_lib}{handler lib name}

\item{show_all_funs}{Default is TRUE and to show all functions in the handler_lib}

\item{show_code}{Default is NULL, select handler you want to see its source code}

\item{show_description}{Default is FALSE}

\item{handler_confg_file}{ngstk handler configuration file path, default is
system.file('extdata', 'config/handler.toml', package = 'ngstk')}
}
\description{
Function to show all avaliabe handler function
}
\examples{
show_handlers(show_description = TRUE)
show_handlers(show_description = FALSE, show_all_funs = FALSE, 
            show_code = 'handler_na_replace')
}
