% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{show_mhandlers}
\alias{show_mhandlers}
\title{Function to show all avaliabe mhandler function}
\usage{
show_mhandlers(mhandler_lib = "default_mhandlers", show_all_funs = TRUE,
  show_code = NULL, show_description = FALSE,
  mhandler_confg_file = system.file("extdata", "config/mhandler.toml", package
  = "ngstk"))
}
\arguments{
\item{mhandler_lib}{handler lib name}

\item{show_all_funs}{Default is TRUE and to show all functions in the handler_lib}

\item{show_code}{Default is NULL, select handler you want to see its source code}

\item{show_description}{Default is FALSE}

\item{mhandler_confg_file}{ngstk handler configuration file path, default is 
system.file('extdata', 'config/handler.toml', package = 'ngstk')}
}
\description{
Function to show all avaliabe mhandler function
}
\examples{
show_mhandlers(show_description = TRUE)
show_mhandlers(show_description = FALSE, show_all_funs = FALSE, 
             show_code = 'mhandler_fusions_anyfull_match')
}
