% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_col_data}
\alias{split_col_data}
\title{Data split function by colum}
\usage{
split_col_data(x, sections = 1)
}
\arguments{
\item{x}{Data.frame or data.table object that need to be divided n sections by colum}

\item{sections}{Split section number (colnum)}
}
\description{
Data split function by colum
}
\examples{
x1 <- data.frame(col1 = 1:10, col2 = 11:20)
x1.t <- t(x1)
x <- split_col_data(x1.t, sections = 3)
}
