% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{supress_any_message}
\alias{supress_any_message}
\title{suppressWarnings(suppressMessages(...))}
\usage{
supress_any_message(...)
}
\arguments{
\item{...}{Paramerters pass to downstream functions}
}
\description{
suppressWarnings(suppressMessages(...))
}
\examples{
supress_any_message(ls())
}
