% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stamp.R
\name{time_stamp}
\alias{time_stamp}
\title{Function to generate time stamp in the files name or directories name.}
\usage{
time_stamp(template = c("\%Y_\%m_\%d_\%H_\%M_\%S_",
  "\%Y_\%m_\%d_\%H_\%M_", "\%Y_\%m_\%d_\%H_", "\%Y_\%m_\%d_", "\%Y_\%m_",
  "\%Y_"), extra_flag = c("-", "/", "@"))
}
\arguments{
\item{template}{The template time format with '_'}

\item{extra_flag}{Using extra flag to replace template '_' and get extra time stamp ['-', '-']}
}
\value{
A list
}
\description{
Function to generate time stamp in the files name or directories name.
}
\examples{
time_stamp()
}
