% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinpaint.R
\name{fusion2pp}
\alias{fusion2pp}
\title{Function to convert fusion data to ProteinPaint \url{https://pecan.stjude.org/pp} input format.}
\usage{
fusion2pp(input_data, input_type = "fusioncatcher",
  config_file = system.file("extdata", "config/proteinpaint.toml", package =
  "ngstk"))
}
\arguments{
\item{input_data}{A gene fusions data.frame need to be converted to ProteinPaint input.}

\item{input_type}{Point the input data format (fusioncatcher or others)}

\item{config_file}{ngstk ProteinPaint configuration file path, default is 
system.file("extdata", "config/proteinpaint.toml", package = "ngstk")}
}
\value{
A data frame
}
\description{
Function to convert fusion data to ProteinPaint \url{https://pecan.stjude.org/pp} input format.
}
\examples{
example <- NULL
}
