% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinpaint_transefer.R
\name{fusion2pp_meta}
\alias{fusion2pp_meta}
\title{Function to convert fusion data to ProteinPaint heatmap meta rows \url{https://pecan.stjude.org/pp} input format.}
\usage{
fusion2pp_meta(input_data, input_type = "fusioncatcher",
  config_file = system.file("extdata", "config/proteinpaint.toml", package =
  "ngstk"))
}
\arguments{
\item{input_data}{A gene fusions data.frame need to be converted to ProteinPaint input.}

\item{input_type}{Point the input data format (fusioncatcher or others)}

\item{config_file}{ngstk ProteinPaint configuration file path, default is}
}
\value{
A data frame
}
\description{
Function to convert fusion data to ProteinPaint heatmap meta rows \url{https://pecan.stjude.org/pp} input format.
}
\examples{
demo_file <- system.file('extdata', 
'demo/proteinpaint/fusion2pp_fusioncatcher.txt', package = 'ngstk')
input_data <- read.table(demo_file, sep = '\\t', header = TRUE, stringsAsFactors = FALSE)
disease <- 'B-ALL'
sampletype <- 'diagnose'
input_data <- data.frame(input_data, disease, sampletype)
input_data$disease <- as.character(input_data$disease)
result <- fusion2pp_meta(input_data, input_type = 'fusioncatcher')
}
