% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngstk.R
\docType{package}
\name{ngstk}
\alias{ngstk}
\alias{ngstk-package}
\title{ngstk can be used to facilitate the analysis of NGS data,
such as visualization, conversion of data format for WEB service input and other purpose.}
\description{
ngstk can be used to facilitate the analysis of NGS data,
such as visualization, conversion of data format for WEB service input and other purpose.
}
\seealso{
Useful links:

\url{https://github.com/JhuangLab/ngstk}

Report bugs at \url{https://github.com/JhuangLab/ngstk/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
